feat(get_type_hierarchy): further work
